Hero page refactor
