+++ src/api/resultsData.js
@@ -1,0 +1,105 @@
+export const SORT_OPTIONS = [
+    {
+        id: 1,
+        label: "Release date",
+        value: 0,
+        isDefault: true,
+    },
+    {
+        id: 2,
+        label: "Year",
+        value: 1,
+        isDefault: false,
+    },
+    {
+        id: 3,
+        label: "Movie name",
+        value: 2,
+        isDefault: false,
+    },
+];
+
+export const FILTER_DATA = [
+    {
+        id: "1",
+        name: "Genre",
+        value: "All",
+        isChecked: true,
+    },
+    {
+        id: "2",
+        name: "Genre",
+        value: "Documentary",
+        isChecked: false,
+    },
+    {
+        id: "3",
+        name: "Genre",
+        value: "Comedy",
+        isChecked: false,
+    },
+    {
+        id: "4",
+        name: "Genre",
+        value: "Horror",
+        isChecked: false,
+    },
+    {
+        id: "5",
+        name: "Genre",
+        value: "Crime",
+        isChecked: false,
+    },
+];
+
+export const RESULTS_COUNT = 39;
+export const RESULTS = [
+  {
+    id: "1",
+    title: "Pulp Fiction",
+    genres: ["Action & Adventure"],
+    production_year: "2004",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+  {
+    id: "2",
+    title: "Bohemian Rhapsody",
+    genres: ["Drama", "Biography", "Music"],
+    production_year: "2003",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+  {
+    id: "3",
+    title: "Kill Bill",
+    genres: ["Oscar winning movie"],
+    production_year: "1994",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+  {
+    id: "4",
+    title: "Avengers: War of Inifinity",
+    genres: ["Action & Adventure"],
+    production_year: "2004",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+  {
+    id: "5",
+    title: "Inception",
+    genres: ["Action & Adventure"],
+    production_year: "2003",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+  {
+    id: "6",
+    title: "Reservoir dogs: ",
+    genres: ["Oscar winning movie"],
+    production_year: "1994",
+    poster_url:
+      "https://s.studiobinder.com/wp-content/uploads/2017/12/Movie-Poster-Template-Dark-with-Image.jpg?x81279",
+  },
+];
