+++ src/components/App/App.tsx
@@ -1,15 +1,16 @@
 import React from "react";
-import { Hero } from "../../containers/Hero/Hero";
-import { Results } from "../../containers/Results/Results";
-import { ResultsModifiers } from "../../containers/ResultsModifiers/ResultsModifiers";
+import ErrorBoundary from "../../hoc/ErrorBoundary/ErrorBoundary";
+import { Hero } from "../Hero/Hero";
+import { Results } from "../Results/Results";
 import "./App.scss";
 
 export const App = () => {
   return (
     <div className="app">
-      <Hero />
-      <ResultsModifiers />
-      <Results />
+      <ErrorBoundary>
+        <Hero />
+        <Results />
+      </ErrorBoundary>
     </div>
   );
 };
