+++ src/components/Movie/MovieCard/MovieCard.types.ts
@@ -1,0 +1,7 @@
+export interface IMovieCardProps {
+    id: string,
+    title: string,
+    poster_url: string,
+    production_year: string,
+    genres: string[] | string
+}
