+++ src/components/Movie/MovieContextMenu/MovieContextMenu.scss
@@ -1,0 +1,15 @@
+@use '../../../styles/colors';
+
+.context-menu {
+    position: absolute;
+    right: 0;
+    padding: .5rem .5rem;
+
+    &__options {
+        display: flex;
+        flex-direction: column;
+        position: absolute;
+        z-index: 10;
+        background-color: colors.$light-grey;
+    }
+}
