+++ src/components/Movie/MovieList/MovieList.scss
@@ -1,0 +1,9 @@
+.movies {
+  display: grid;
+  align-items: center;
+  grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
+  grid-gap : 1rem;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
