+++ src/components/Movie/MovieList/MovieList.types.ts
@@ -1,0 +1,3 @@
+import { IMovieCardProps } from "../MovieCard/MovieCard.types";
+
+export interface IMovieListProps extends Array<IMovieCardProps> {}
