--- src/components/MovieCard/MovieCard.types.ts
@@ -1,7 +1,0 @@
-export interface IMovieCardProps {
-    id: string,
-    title: string,
-    posterUrl: string,
-    productionYear: string,
-    genres: string[] | string
-}
