--- src/components/MovieContextMenu/MovieContextMenu.scss
@@ -1,12 +1,0 @@
-.context-menu {
-    position: absolute;
-    right: 0;
-    padding: .5rem .5rem;
-
-    &__options {
-        display: flex;
-        flex-direction: column;
-        position: absolute;
-        z-index: 10;
-    }
-}
