+++ src/components/Results/Filter/Filter.scss
@@ -1,0 +1,29 @@
+@use "../../../styles/colors";
+
+.filter {
+  $this: &;
+
+  &__input {
+    opacity: 0;
+    position: fixed;
+    width: 0;
+
+    &:checked + #{$this}__label:after {
+      content: "";
+      height: 0.2rem;
+      background-color: colors.$pastel-red;
+      position: absolute;
+      left: 0;
+      bottom: 0;
+      width: 100%;
+      z-index: 10;
+    }
+  }
+
+  &__label {
+    display: inline-block;
+    padding: 0.5rem 0.8rem;
+    position: relative;
+    text-transform: uppercase;
+  }
+}
