+++ src/components/Results/Filter/Filter.types.ts
@@ -1,0 +1,6 @@
+export interface IFilterProps {
+      id: string,
+      name: string,
+      value: string,
+      isChecked: boolean
+}
