+++ src/components/Results/Filter/FilterList/FilterList.scss
@@ -1,0 +1,5 @@
+.filters {
+  display: flex;
+  align-items: center;
+  overflow: auto;
+}
