+++ src/components/Results/Filter/FilterList/FilterList.types.ts
@@ -1,0 +1,3 @@
+import { IFilterProps } from "../Filter.types"
+
+export interface IFilterListProps extends Array<IFilterProps> {}
