+++ src/components/Results/Results.types.ts
