+++ src/components/Results/ResultsCount/ResultsCount.types.ts
@@ -1,0 +1,3 @@
+export interface IPropsResultsCount {
+  value: number
+}
