+++ src/components/Results/Sort/Sort.scss
@@ -1,0 +1,37 @@
+@use "../../../styles/colors";
+
+.sort {
+  display: flex;
+  align-items: center;
+  font-size: inherit;
+  text-transform: uppercase;
+  gap: 0 1rem;
+  position: relative;
+  font-size: 0.9rem;
+
+  &::after {
+    content: "\25BC";
+    position: absolute;
+    right: -1rem;
+    color: colors.$pastel-red;
+  }
+
+  &__label {
+    color: colors.$alt-grey;
+    letter-spacing: 0.1ch;
+    white-space: nowrap;
+  }
+
+  &__input {
+    background: transparent;
+    border: none;
+    -webkit-appearance: none;
+    appearance: none;
+    color: colors.$white;
+    font-size: inherit;
+    font-family: inherit;
+    font-weight: 700;
+    letter-spacing: 0.1ch;
+    text-transform: uppercase;
+  }
+}
