+++ src/components/Results/Sort/Sort.types.ts
@@ -1,0 +1,10 @@
+export interface ISortProps {
+    options: Array<{
+        id: number,
+        label: string,
+        value: string | number,
+        isDefault?: boolean
+    }>,
+    onOptionChange: () => void
+    sortLabel: string
+}
