+++ src/components/Search/Search.scss
@@ -1,0 +1,24 @@
+@use "./../../styles/colors";
+@use "./../../styles/variables";
+
+.search {
+  display: flex;
+
+  > input {
+    font-size: inherit;
+    font-family: inherit;
+    color: colors.$white;
+    background-color: colors.$dark;
+    border: 1px solid colors.$dark-grey;
+    border-radius: 3px;
+    padding: variables.$section-padding;
+    height: 100%;
+    width: 100%;
+    margin-right: .5rem;
+  }
+
+  ::placeholder {
+    color: inherit;
+    font-family: inherit;
+  }
+}
