+++ src/components/Search/Search.types.ts
@@ -1,0 +1,6 @@
+export interface ISearchProps {
+    placeholder?: string,
+    searchButtonText?: string
+    setQuery: () => void;
+    onSearch: () => void;
+}
