--- src/components/UI/Filter/Filters.scss
@@ -1,52 +1,0 @@
-@use "../../../styles/colors";
-
-// .filter {
-//     &__input {
-//         position: absolute;
-//         margin: 0 !important;
-//         padding: 0 !important;
-//         opacity: 0;
-//         height: 0;
-//         width: 0;
-//         pointer-events: none;
-//     }
-// }
-
-.filters {
-  display: flex;
-  align-items: center;
-  overflow: auto;
-}
-
-.filter {
-  $this: &;
-  //   scroll-snap-align: start;
-
-  &__input {
-    opacity: 0;
-    position: fixed;
-    width: 0;
-
-    &:checked + #{$this}__label:after {
-      content: "";
-      height: 0.2rem;
-      background-color: colors.$pastel-red;
-      position: absolute;
-      left: 0;
-      bottom: 0;
-      width: 100%;
-      z-index: 10;
-    }
-  }
-
-  &__label {
-    display: inline-block;
-    padding: 0.5rem 0.8rem;
-    position: relative;
-    text-transform: uppercase;
-
-    // .filters > #{$this}:first-of-type > & {
-    //   padding-left: 0;
-    // }
-  }
-}
