--- src/components/UI/Filter/Filters.types.ts
@@ -1,8 +1,0 @@
-export interface IFiltersProps {
-    data: Array<{
-        id: string,
-        name: string,
-        value: string,
-        isChecked: boolean
-    }>
-}
