--- src/components/UI/SearchBar/SearchBar.types.ts
@@ -1,4 +1,0 @@
-export interface ISearchBarProps {
-    placeholder?: string,
-    setQuery: () => void;
-}
