--- src/components/UI/Select/Select.types.ts
@@ -1,10 +1,0 @@
-export interface ISelectProps {
-    options: Array<{
-        id: number,
-        label: string,
-        value: string | number,
-        isDefault?: boolean
-    }>,
-    onOptionChange: () => void
-    selectLabel: string
-}
