--- src/containers/Results/Results.scss
@@ -1,23 +1,0 @@
-@use "../../styles/variables";
-
-.results {
-  margin: 0;
-  max-width: variables.$section-max-width;
-  padding: variables.$section-padding;
-
-  &-number {
-    display: inline-block;
-    padding-left: 0.7rem;
-
-    &--bold {
-      font-weight: bold;
-    }
-  }
-}
-
-.movies {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 2rem;
-  padding: variables.$section-padding;
-}
