--- src/containers/ResultsModifiers/randomdata.js
@@ -1,53 +1,0 @@
-export const SORT_OPTIONS = [
-    {
-        id: 1,
-        label: "Release date",
-        value: 0,
-        isDefault: true,
-    },
-    {
-        id: 2,
-        label: "Year",
-        value: 1,
-        isDefault: false,
-    },
-    {
-        id: 3,
-        label: "Movie name",
-        value: 2,
-        isDefault: false,
-    },
-];
-
-export const FILTER_DATA = [
-    {
-        id: "1",
-        name: "Genre",
-        value: "All",
-        isChecked: true,
-    },
-    {
-        id: "2",
-        name: "Genre",
-        value: "Documentary",
-        isChecked: false,
-    },
-    {
-        id: "3",
-        name: "Genre",
-        value: "Comedy",
-        isChecked: false,
-    },
-    {
-        id: "4",
-        name: "Genre",
-        value: "Horror",
-        isChecked: false,
-    },
-    {
-        id: "5",
-        name: "Genre",
-        value: "Crime",
-        isChecked: false,
-    },
-];
