+++ src/hoc/ErrorBoundary/ErrorBoundary.type.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface IPropsErrorBoundary {
+  children: ReactNode;
+}
